+++ src/utils/src/Coroutine.php
@@ -42,15 +42,16 @@ class Coroutine
 
     /**
      * Returns the parent coroutine ID.
-     * Returns -1 when running in non-coroutine context.
+     * Returns -1 when running in the top level coroutine.
+     * Returns null when running in non-coroutine context.
      *
      * @see https://github.com/swoole/swoole-src/pull/2669/files#diff-3bdf726b0ac53be7e274b60d59e6ec80R940
      */
-    public static function parentId(): int
+    public static function parentId(): ?int
     {
         $cid = SwooleCoroutine::getPcid();
         if ($cid === false) {
-            return -1;
+            return null;
         }
 
         return $cid;